use one config js file
