--- src/js/replay.js
@@ -1,11 +1,0 @@
-/** Replay config */
-import config from '../wrg-config.json';
-
-class ReplayOptions {
-  constructor({ embed, replayBase }) {
-    this.embed = embed || 'embed';
-    this.replayBase = replayBase || './replay/';
-  }
-}
-
-export default new ReplayOptions(config.replay);
